Modifications in POM classes since rework in create / join / start chat buttons
- For Android & iOS POM classes
