+++ VectorMobileTests/src/test/java/pom_ios/main_tabs/RiotPeopleTabPageObjects.java
@@ -47,18 +47,4 @@ public class RiotPeopleTabPageObjects extends RiotTabPageObjects{
 	public MobileElement getPeopleByName(String peopleDisplayName){
 		return null;
 	}
-	/*
-	 * FLOATING BUTTONS OR DIALOGS .
-	 */
-	/**
-	 * Start a new chat with a user and returns the new created room. 
-	 * @param displayNameOrMatrixId
-	 * @return RiotRoomPageObjects
-	 */
-	public RiotRoomPageObjects startChat(String displayNameOrMatrixId){
-		createRoomFloatingButton.click();
-		RiotNewChatPageObjects newChatA = new RiotNewChatPageObjects(appiumFactory.getiOsDriver1());
-		newChatA.searchAndSelectMember(displayNameOrMatrixId);
-		return new RiotRoomPageObjects(driver);
-	}
 }
